pyproject.toml
[project]
name = "multimodal-diffusion-policies"
version = "0.1.0"
description = "Multimodal diffusion policies for robotic manipulation with vision and tactile sensing"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["robotics", "diffusion models", "imitation learning", "tactile sensing", "pytorch"]

# ---------------------------------------------------------------------------
# Core dependencies
#
# NOTE: CleanDiffuser is a required external dependency that must be installed
# manually before this package, as it is not available on PyPI:
#
#   git clone https://github.com/CleanDiffuserTeam/CleanDiffuser.git
#   pip install -e CleanDiffuser/
#
# See README.md for full setup instructions.
# ---------------------------------------------------------------------------
dependencies = [
    # deep learning
    "torch>1.0.0,<2.3.0",
    "torchvision",
    "numpy<1.23.0",
    "einops",
    "timm==1.0.20",

    # data
    "h5py==3.14.0",
    "zarr>=2.16.1,<2.17",
    "imagecodecs",

    # config & logging
    "hydra-core",
    "wandb",
    "omegaconf",

    # server / client
    "fastapi==0.118.0",
    "uvicorn==0.37.0",
    "draccus==0.11.5",
    "json-numpy==2.1.1",
    "requests",

    # vision & utilities
    "scipy",
    "scikit-image<0.23.0",
    "opencv-python",
    "matplotlib",
    "Pillow",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["source*"]
